requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.0
alembic>=1.10.0
psycopg2-binary>=2.9.5
python-jose>=3.3.0
passlib>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6
email-validator>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
tarantool>=0.9.0
python-dotenv>=1.0.0
pillow>=9.5.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
pytest-cov>=4.0.0